Isolated event (regular and recommended) fetching and page count fetching logic into a util file
